Downloads/ISA-hourly-performance-main - Copy/ISA-hourly-performance-main/requirements.txt
autopep8==2.0.2
black==23.1.0
click==8.1.3
colorama==0.4.6
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
